backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
onnxruntime>=1.16.0
tokenizers>=0.15.0
numpy>=1.24.0
requests>=2.31.0
qdrant-client>=1.6.0
faiss-cpu>=1.7.4
httpx>=0.25.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
rank-bm25>=0.2.2
gitpython>=3.1.40
gunicorn>=21.2.0
groq
python-dotenv
